+++ axelor-sale/src/main/java/com/axelor/apps/sale/web/OpportunitySaleOrderController.java
@@ -21,6 +21,7 @@ import com.axelor.apps.crm.db.Opportunity;
 import com.axelor.apps.crm.db.repo.OpportunityRepository;
 import com.axelor.apps.sale.db.SaleOrder;
 import com.axelor.apps.sale.service.saleorder.OpportunitySaleOrderService;
+import com.axelor.apps.sale.service.saleorder.SaleOrderWorkflowService;
 import com.axelor.exception.AxelorException;
 import com.axelor.inject.Beans;
 import com.axelor.meta.schema.actions.ActionView;
@@ -28,11 +29,13 @@ import com.axelor.rpc.ActionRequest;
 import com.axelor.rpc.ActionResponse;
 import com.google.inject.Inject;
 import com.google.inject.Singleton;
+import java.util.List;
 
 @Singleton
 public class OpportunitySaleOrderController {
 
   @Inject private OpportunitySaleOrderService opportunitySaleOrderService;
+  @Inject private SaleOrderWorkflowService saleOrderWorkflowService;
 
   public void generateSaleOrder(ActionRequest request, ActionResponse response)
       throws AxelorException {
@@ -47,5 +50,15 @@ public class OpportunitySaleOrderController {
             .param("forceEdit", "true")
             .context("_showRecord", String.valueOf(saleOrder.getId()))
             .map());
+  }
+
+  public void cancelSaleOrders(ActionRequest request, ActionResponse response) {
+    Opportunity opportunity = request.getContext().asType(Opportunity.class);
+    List<SaleOrder> saleOrderList = opportunity.getSaleOrderList();
+    if (saleOrderList != null && !saleOrderList.isEmpty()) {
+      for (SaleOrder saleOrder : saleOrderList) {
+        saleOrderWorkflowService.cancelSaleOrder(saleOrder, null, opportunity.getName());
+      }
+    }
   }
 }